Many updates and fixes...
Pick rates had a major bug. Ascension level and floor filters were not the same for pick counts and not-picked counts. This caused calculations to be very wrong. This explains why offer rates seemed drastically different for different cards of the same rarity.

Correlations page now does an export which I used for the web app.